% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{length of sample size}

\item{fun}{function that will be applied to xstat vector}

\item{alpha}{alpha value to determine percentage of confidence interval}

\item{cx}{numeric character expansion factor}

\item{...}{extra param}
}
\value{
histogram of bootstrap sample statistics and list(ci=ci,fun=fun,x=x,xstat=xstat)
}
\description{
Title
}
