% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Title}
\usage{
myci(x)
}
\arguments{
\item{x}{input sample}
}
\value{
95 percent confidence interval
}
\description{
Title
}
\examples{
d=c(5.0581, 4.9707, 5.0893, 4.9334, 4.9777, 5.0285, 4.8555, 4.9565)
myci(d)
}
