% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Title}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample size (for each iteration)}

\item{iter}{Number of iterations}

\item{a}{Lower limit of the sample from the uniform distribution}

\item{b}{Upper limit of the sample from the uniform distribution}
}
\value{
A histogram of the sample means generated from a uniform distribution. Also has a theoretical normal curve and density curve
}
\description{
Title
}
\examples{
mycltu(n=20,iter=100000)
}
