% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Title}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A likelihood function that takes a single parameter and returns the log-likelihood value. This function is applied to each value of `theta`. It should be provided as a function name without quotes.}

\item{theta}{A numeric vector of possible parameter values to evaluate, representing the range of values for the parameter of interest.}
}
\value{
The function returns the value of `theta` that maximizes the likelihood.
}
\description{
Title
}
\examples{
# Example usage with a custom likelihood function
logbin2 <- function(theta) { sum(dbinom(c(3,3,4,3,4,5,5,4), size=20, prob=theta, log=TRUE)) }
mymaxlikg(lfun=logbin2, theta=seq(0, 1, length=10000))
}
