% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The upper bound of the lower tail}
}
\value{
A plot of the normal distribution with polygon that shades the lower tail and a output of the probability rounded to 4 decimal place.
}
\description{
myncurve
}
\examples{
myncurve(10, 5, 6)
}
