% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Title}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc value}

\item{xmax}{bound of the x-axis}

\item{n}{Number of values in the sample}

\item{alpha}{alpha value (if 0.05 then 95 percent C.I.)}
}
\value{
plot of normal curve and quantiles, pvalue and q value
}
\description{
Title
}
\examples{
tcalc = 1.326252
mypvalue(tcalc,n=30,alpha=0.05)
}
