% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Title}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{The size of the samples}

\item{iter}{The number of iterations}

\item{time}{Pause between each iteration (Seconds)}
}
\value{
Returns a Barplot for each iteration of random sampling with replace being True
}
\description{
Title
}
\examples{
mysample(n=1000, iter=3, time=1)
}
