% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{nticket}
\alias{nticket}
\title{Title}
\usage{
nticket(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{The total number of seats in the plane}

\item{gamma}{The probability a plane will be truly overbooked}

\item{p}{The probability of a show}
}
\value{
Return a named list with number of tickets calculated with discrete distribution, number of tickets calculated with normal approximation, N, p, and gamma
}
\description{
Title
}
\examples{
nticket(N=400,gamma = 0.02, p = 0.95)
}
