% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histo}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X Variable as a vector}

\item{y}{Y Variable as a vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A sophisticated plot (ScatterHist)
}
\description{
Scatter Histo
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
